LayoutTests/http/tests/security/xssAuditor/full-block-iframe-no-inherit-expected.txt

CONSOLE MESSAGE: Refused to execute a JavaScript script. Source code of script found within request.

This tests that the header X-XSS-Protection is not inherited by the iframe below:



--------
Frame: 'frame'
--------
If you see this message and no JavaScript alert() then the test PASSED.
